feat"Softskills satisfaction , edit cohort , assign trainers into cohorts" adit" fixed the loggin page"

+++ database/migrations/2024_03_29_192716_create_soft_skills_trainings_table.php
@@ -19,7 +19,7 @@ class CreateSoftSkillsTrainingsTable extends Migration
             $table->text('description');
             $table->string('trainer');
             $table->date('date');
-            $table->integer('satisfaction')->nullable();
+            $table->double('satisfaction', 8, 2)->nullable();
             $table->unsignedBigInteger('cohort_id');
             $table->foreign('cohort_id')->references('id')->on('cohorts')->onDelete('cascade');
             $table->timestamps();
